fix(Decoder): Correct integer-division bug in view-split transform gain
PartitionDecoder::transformGain re-derives the constant gain the
encoder applies to every leaf (TransformPartition::totalTransformGain,
a fixed sqrt(product of the top-level block's dimensions)) from the
leaf's split length and the full partition size. The division
mPartitionData.data.size(i)/length[i] was done in int64_t, truncating
whenever a split doesn't divide the full size evenly.

Spatial (v,u) block sizes are chosen as powers of two, so they always
divide evenly and the bug was invisible there. View (t,s) sizes are
not (e.g. 9 splits into 4/5), so an inter-view split leaf got a wrong,
and different-per-quadrant, gain applied during reconstruction. Since
this scales the whole leaf uniformly and DC dominates block energy,
the visible symptom was decoded blocks with the correct shape/position
but visibly wrong brightness at inter-view split boundaries.

Verified with a minimal single-block repro (view split 9->4/5): before
the fix the four split quadrants decoded with gains 64, 50.6, 40, 50.6
instead of the encoder's constant 72; after the fix all four match.

+++ src/Decoder/PartitionDecoder.cpp
@@ -104,7 +104,7 @@ double PartitionDecoder :: transformGain(std::array<int64_t,4> length){
     for(int i = 0; i < 4; i++){
         transformGain*=length[i]/sqrt(length[i]);
 
-        transformGain  *= sqrt(mPartitionData.data.size(i)/length[i]);
+        transformGain  *= sqrt((double)mPartitionData.data.size(i)/(double)length[i]);
     }
     return transformGain*mGain;
 }
